fix(arc/cpp): update stale channel/ includes after rename to channels/
